Refactor provider and context file structure - move context to its own provider for better reusability - lessen complexity
