+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { AppContext, AppProvider as Provider } from './providers';
+import { AppContext, AppProvider } from './providers/AppProvider';
 import React, { useContext } from 'react';
 import { hydrate, render } from 'react-dom';
 
@@ -7,9 +7,9 @@ import SwitchNavigator from './components/navigation/SwitchNavigator';
 const rootElement = document.getElementById('app') as HTMLElement;
 
 const Component = () => (
-  <Provider>
+  <AppProvider>
     <SwitchNavigator />
-  </Provider>
+  </AppProvider>
 );
 
 const renderApp = () => {
