--- src/providers/index.ts
@@ -1,7 +1,0 @@
-import { AppConsumer, AppContext, AppProvider } from './AppProvider';
-
-export {
-  AppProvider,
-  AppConsumer,
-  AppContext,
-};
